Separate out paginator function
