+++ templates/createPosts.js
@@ -1,12 +1,13 @@
 const path = require("path")
-const { postsPerPage } = require("./config/website")
+const { postsPerPage } = require("../config/website")
+const { paginator } = require("./helpers/paginator")
 
 exports.createPosts = ({ createPage }) => (result) => {
   const postTemplate = path.resolve("src/pages-generic/post.jsx")
   const postsTemplate = path.resolve("src/pages-generic/posts.jsx")
 
   const posts = result.data.prismic.allPosts.edges
-  const totalCount = result.data.prismic.allInsights.totalCount
+  const totalCount = result.data.prismic.allPosts.totalCount
 
   posts.forEach((post) => {
     const uid = post.node._meta.uid
@@ -20,22 +21,11 @@ exports.createPosts = ({ createPage }) => (result) => {
     })
   })
 
-  const numPages = Math.ceil(totalCount / postsPerPage)
-
-  Array.from({ length: numPages }).forEach((_, i) => {
-    const limit = postsPerPage
-    const skip = i * postsPerPage
-    const currentPage = i + 1
-
-    createPage({
-      path: i === 0 ? "/insights" : `/insights/${i + 1}`,
-      component: postsTemplate,
-      context: {
-        limit,
-        skip: limit - skip,
-        numPages,
-        currentPage,
-      },
-    })
+  paginator({
+    totalCount,
+    postsPerPage,
+    path: "insights",
+    template: postsTemplate,
+    createPage,
   })
 }
